Ebpf, Daemon: Share socket info, implement listen socket tracking
The state change for TCP sockets to TCP_LISTEN seems to only happen in `inet_socket_set_state` but the socket cookie is not available there so we now share additional socket info in a map that uses the socket address as the key which seems to be consistent across the functions.

We don't have the local address of the socket for now. Not sure if it can even be retrieved.