IPLink: Basic setup
this creates a new interface for ingress bandwidth shaping as well as the necessary qdiscs, classes and filters.

The basic idea is to use HTB and qdiscs to limit bandwidth. For this to work we need to add a filter and a qdisc for each limit that is to be enforced. The qdisc handles the throttling, the filter redirects the packets to the qdisc.

By default this only works for outgoing traffic, so we redirect incoming traffic to an ifb interface, this way it effectively turns into outgoing traffic which can then also be limited. At a later point this should be done via the netlink library but for now using the tc CLI interface is faster for testing.